requirements.txt
websocket-client
pynput
